+++ web-common/src/proto/gen/rill/runtime/v1/api_pb.ts
@@ -856,6 +856,21 @@ export class CreateInstanceRequest extends Message$1<CreateInstanceRequest> {
    */
   connectors: Connector[] = [];
 
+  /**
+   * @generated from field: map<string, string> variables = 7;
+   */
+  variables: { [key: string]: string } = {};
+
+  /**
+   * @generated from field: map<string, string> annotations = 9;
+   */
+  annotations: { [key: string]: string } = {};
+
+  /**
+   * @generated from field: string frontend_url = 18;
+   */
+  frontendUrl = "";
+
   constructor(data?: PartialMessage<CreateInstanceRequest>) {
     super();
     proto3.util.initPartial(data, this);
@@ -871,6 +886,9 @@ export class CreateInstanceRequest extends Message$1<CreateInstanceRequest> {
     { no: 15, name: "admin_connector", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 17, name: "ai_connector", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 10, name: "connectors", kind: "message", T: Connector, repeated: true },
+    { no: 7, name: "variables", kind: "map", K: 9 /* ScalarType.STRING */, V: {kind: "scalar", T: 9 /* ScalarType.STRING */} },
+    { no: 9, name: "annotations", kind: "map", K: 9 /* ScalarType.STRING */, V: {kind: "scalar", T: 9 /* ScalarType.STRING */} },
+    { no: 18, name: "frontend_url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): CreateInstanceRequest {
