+++ web-common/src/runtime-client/gen/index.schemas.ts
@@ -764,6 +764,10 @@ export interface V1CreateDirectoryResponse {
   [key: string]: unknown;
 }
 
+export type V1CreateInstanceRequestVariables = { [key: string]: string };
+
+export type V1CreateInstanceRequestAnnotations = { [key: string]: string };
+
 /**
  * Request message for RuntimeService.CreateInstance.
 See message Instance for field descriptions.
@@ -776,6 +780,9 @@ export interface V1CreateInstanceRequest {
   adminConnector?: string;
   aiConnector?: string;
   connectors?: V1Connector[];
+  variables?: V1CreateInstanceRequestVariables;
+  annotations?: V1CreateInstanceRequestAnnotations;
+  frontendUrl?: string;
 }
 
 export interface V1CreateInstanceResponse {
